create task

+++ bot/handlers/tasks/add.py
@@ -9,7 +9,9 @@ from aiogram_calendar import SimpleCalendar, simple_cal_callback
 from bot.commands import CommandEnum
 from bot.keyboards.categories import create_retrieve_category_keyboard
 from bot.loader import bot, dp
+from bot.schemes.tasks import TaskCreateScheme
 from bot.services.categories import CategoryService
+from bot.services.tasks import TaskService
 from bot.states import FSMAddTask
 
 
@@ -50,8 +52,10 @@ async def add_planed_time(message: types.Message, state: FSMContext) -> None:
 @dp.callback_query_handler(simple_cal_callback.filter(), state=FSMAddTask.planed_time)
 async def add_timeshift(callback_query: types.CallbackQuery, state: FSMContext, callback_data: Dict) -> None:
     selected, date = await SimpleCalendar().process_selection(callback_query, callback_data)
+    if not selected:
+        return
     async with state.proxy() as data:
-        data["planed_time"] = date.strftime("%d/%m/%Y")
+        data["planed_time"] = date
     await bot.send_message(callback_query.from_user.id, "Укажите время на выполнение задачи")
     await FSMAddTask.next()
     await bot.delete_message(callback_query.from_user.id, callback_query.message.message_id)
@@ -64,15 +68,16 @@ async def add_links(message: types.Message, state: FSMContext) -> None:
             data["timeshift"] = int(message.text)
         except Exception:
             data["timeshift"] = None
-    await bot.send_message(message.from_user.id, "Укажите возможные ссылки через запятую")
-    await FSMAddTask.next()
-    await bot.delete_message(message.from_user.id, message.message_id)
-
-
-@dp.message_handler(state=FSMAddTask.links)
-async def end_add_task(message: types.Message, state: FSMContext) -> None:
-    async with state.proxy() as data:
-        data["links"] = [link.strip() for link in message.text.split(",")]
+        await TaskService.create(TaskCreateScheme(**data, user_id=message.from_user.id))
         await bot.send_message(message.from_user.id, data)
     await state.finish()
     await bot.delete_message(message.from_user.id, message.message_id)
+
+
+# @dp.message_handler(state=FSMAddTask.links)
+# async def end_add_task(message: types.Message, state: FSMContext) -> None:
+#     async with state.proxy() as data:
+#         data["links"] = [link.strip() for link in message.text.split(",")]
+#         await bot.send_message(message.from_user.id, data)
+#     await state.finish()
+#     await bot.delete_message(message.from_user.id, message.message_id)
